Refactor workflow.py
