+++ hypha/apply/determinations/permissions.py
@@ -1,4 +1,5 @@
-from .options import DETERMINATION_OUTCOMES
+from hypha.apply.funds.workflow import DETERMINATION_OUTCOMES
+
 from .utils import determination_actions, transition_from_outcome
